bit better - evaluation simplified, ordering of possible moves redone

+++ chessGame/PotentialMove.cs
@@ -10,8 +10,8 @@ namespace chessGame
     {
         public Piece p;
         public Cell newCell;
-        public double value;
-        public PotentialMove(double value, Piece p, Cell newCell)
+        public long value;
+        public PotentialMove(long value, Piece p, Cell newCell)
         {
             this.value = value;
             this.p = p;
